requirements.txt
streamlit
google_generativeai
python-dotenv
langchain
PyPDF2
#chromabd
faiss-cpu
langchain_google_genai